Add msbuild configurations for the "LowTrust" builds and refactor/rewrite the new project files and pack script
I'll update the documentation and add back the CI configuration in follow-up commits.

Note that the `dotnet restore` command is configuration dependent but doesn't support the '-c' parameter. Instead you have to use e.g. 'dotnet restore /p:Configuration=Debug-LowTrust'.

The new pack script builds both "editions" of the FParsec NuGet package.
The normal (low-trust) package targets 'netstandard1.6', 'net40-client and 'portable-net45+win8+wp8+wpa81' (for package backwards compatibility).
The "Big-Data-Edition" (non-low-trust) package only targets 'net40-client' and 'net45' and thus is not .NET Core compatible.
Each package includes both FParsec.dll and FParsecCS.dll (because the split into the two DLLs is only an implementation detail and FParsecCS.dll is not meant to be used alone).
The net4* and portable assemblies continue to be strong-named for backwards compatibility, but the netstandard assemblies are not strong-named.