Associate the native memory of the emitted code with the dynamic assembly.
We previously associated the memory with the FParsec assembly.
Because the dynamic assembly does not depend on FParsec, it is possible for the former to outlive the latter, which would cause the memory to be swept under its feet.
The memory is now associated with a dummy type we create in the dynamic assembly.
Since the dynamic assembly is not created with RunAndCollect we could have just allocated the memory with a malloc, but it can still be unloaded if it belongs in an unloadable AssemblyLoadContext.
We could have also used `TypeBuilder.DefineInitializedData` but it would need some refactoring or more allocations, and there are few benefits.